## refactor(ecs/factories/monster): modulariza la creación de monstruos en módulos especializados
### Cambios realizados:

- **cache.py**: contiene la lógica de carga y almacenamiento de sprites y death images (`_SPRITE_SURFACES`, `_DEATH_IMAGES`).
- **sprite_loader.py**: implementado `create_sprite_component` y `create_patrol_components` para separar la lógica de animaciones y patrullas.
- **physics.py**: movidas funciones para cálculo de posición y componentes de física:
  - `calculate_position`
  - `create_physics_components`
  - `create_collider_components`
  - `create_zlayer_component`
- **monster_factory.py**: simplificado a una función `spawn_monster` que orquesta las llamadas a los módulos anteriores.

### Objetivo:
Eliminar código inline y distribuir responsabilidades para mejorar la mantenibilidad y reutilización del código.

### Resultado:
Código de creación de monstruos más limpio, modular y testable.
