+++ src/roguelike_game/ecs/core/spawn_manager.py
@@ -3,7 +3,9 @@
 from roguelike_engine.map.utils import calculate_lobby_offset
 from roguelike_engine.config.map_config import global_map_settings
 from roguelike_game.ecs.utils.collider_utils import build_collider_rect
-from roguelike_game.ecs.factories.entity_factory import _load_caches_once, _DEFS, _create_sprite_component, _calculate_position, _create_collider_components
+from roguelike_game.ecs.factories.monster.monster_factory import _load_caches_once, create_sprite_component, calculate_position, create_collider_components
+from roguelike_game.ecs.factories.monster.sprite_loader import create_sprite_component
+from roguelike_game.ecs.factories.monster.config import MONSTER_DEFS
 from roguelike_game.ecs.utils.spawn_utils import find_spawn_positions
 from roguelike_game.ecs.components.spawn.spawn_request import SpawnRequest
 
@@ -27,16 +29,16 @@ class SpawnNPCManager:
 
         # 2) Cargar definiciones y sprite base
         _load_caches_once()
-        cfg = _DEFS["barbol"]
-        sprite, _ = _create_sprite_component("barbol")
+        cfg = MONSTER_DEFS["barbol"]
+        sprite, _ = create_sprite_component("barbol")
         spawned_rects = []
 
         # 3) Spawn en LOBBY
         positions = find_spawn_positions(self.map_manager, self.buildings, lobby_offset, zone_size, neighbor_padding=3, sample_count=10)
         filtered_positions = []
         for tx, ty in positions:
-            px, py = _calculate_position(tx, ty, cfg, sprite)
-            multi = _create_collider_components(sprite, cfg)
+            px, py = calculate_position(tx, ty, cfg, sprite)
+            multi = create_collider_components(sprite, cfg)
             feet = multi.colliders.get("feet")
             if feet:
                 rect = build_collider_rect(px, py, feet)
@@ -58,8 +60,8 @@ class SpawnNPCManager:
         empty_positions = find_spawn_positions(self.map_manager, self.buildings, empty_offset, zone_size, neighbor_padding=3, sample_count=100)
         filtered_empty = []
         for tx, ty in empty_positions:
-            px, py = _calculate_position(tx, ty, cfg, sprite)
-            multi = _create_collider_components(sprite, cfg)
+            px, py = calculate_position(tx, ty, cfg, sprite)
+            multi = create_collider_components(sprite, cfg)
             feet = multi.colliders.get("feet")
             if feet:
                 rect = build_collider_rect(px, py, feet)
